code/requirements.txt
numpy>=1.26,<2.0
opencv-python>=4.8
Pillow>=9.5,<10.0
pytest>=8.0
PySide6>=6.6
torch
torchvision
simple-lama-inpainting>=0.1.2